Move the component test setup file to src/renderer
