+++ app/tests/renderer/views/Inbox.test.tsx
@@ -3,7 +3,7 @@ import { screen } from "@testing-library/react";
 import { expect } from "vitest";
 import userEvent from "@testing-library/user-event";
 import InboxView from "../../../src/renderer/views/Inbox";
-import { renderWithProviders } from "../../component-setup";
+import { renderWithProviders } from "../../../src/renderer/test-component-setup";
 
 describe("InboxView Component", () => {
   it('says the string "Hello world!"', () => {
